+++ odbc-api/src/buffers/text_row_set.rs
@@ -254,8 +254,8 @@ unsafe impl RowSetBuffer for TextRowSet {
 }
 
 unsafe impl ParameterCollection for &TextRowSet {
-    fn parameter_set_size(&self) -> u32 {
-        *self.num_rows as u32
+    fn parameter_set_size(&self) -> usize {
+        *self.num_rows
     }
 
     unsafe fn bind_parameters_to(self, stmt: &mut StatementImpl<'_>) -> Result<(), Error> {